links: added s3 integration
