+++ pkg/interface/src/views/apps/links/components/links-list.js
@@ -137,7 +137,7 @@ export class Links extends Component {
         <div className="w-100 mt6 flex justify-center overflow-y-scroll ph4 pb4">
           <div className="w-100 mw7">
             <div className="flex">
-              <LinkSubmit resourcePath={props.resourcePath} api={this.props.api} />
+              <LinkSubmit resourcePath={props.resourcePath} api={this.props.api} s3={props.s3} />
             </div>
             <div className="pb4">
             {LinkList}
